+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Revision history for th-abstraction
 
+## next -- ????.??.??
+* Fix a bug in which `normalizeDec` would return incorrect results for GADTs
+  that use visible dependent quantification in their return kind.
+
 ## 0.6.0.0 -- 2023.07.31
 * Support building with `template-haskell-2.21.0.0` (GHC 9.8).
 * Adapt to `TyVarBndr`s for type-level declarations changing their type from